Update index.html
Okay, I have a list of changes. First thing is I need it so that when it's shared to desktop on iphone it behaves as a web app or whatever it's called. Early versions had that informaiton, but you removed it. How the days are handled is weird. When I click on a day, Saturday for example, it gives me Saturday night and then sunday AM. I would prefer it do saturday AM and Saturday PM. Also, I want you to store the data somehow, so that when I open the app, you clear all the data until it's refreshed as you currently do, but if after 10 seconds you dont get fresh data, show the old data and put a banner saying when it was last updated. Make that red. Add gust speed to the 24 hour chart. Make it so I can advance that chart 24 hours. Also, it's hard to read the times on that chart. Maybe don't worry about AM and PM and just show the numbers or something to clean it up. Also, when the app opens, load data for the three defaults as well as the last place that was looked at it if it is saved. Also put cloud cover on the 24 hour chart. Let me know if you have any questions please.